refactor: rename use-disclosure file name
