+++ src/hooks/index.js
@@ -1,1 +1,1 @@
-export * from './useDisclosure';
+export * from './use-disclosure';
